bugfix, rename cmd
