+++ cmd/root.go
@@ -6,7 +6,6 @@ import (
 	_ "net/http/pprof" // enable pprof via a flag
 	"os"
 
-	"github.com/bmeg/grip/cmd/caliperload"
 	"github.com/bmeg/grip/cmd/create"
 	"github.com/bmeg/grip/cmd/delete"
 	"github.com/bmeg/grip/cmd/drop"
@@ -14,6 +13,7 @@ import (
 	"github.com/bmeg/grip/cmd/erclient"
 	"github.com/bmeg/grip/cmd/info"
 	"github.com/bmeg/grip/cmd/job"
+	"github.com/bmeg/grip/cmd/jsonload"
 	"github.com/bmeg/grip/cmd/kvload"
 	"github.com/bmeg/grip/cmd/list"
 	"github.com/bmeg/grip/cmd/load"
@@ -61,7 +61,7 @@ func init() {
 	RootCmd.AddCommand(job.Cmd)
 	RootCmd.AddCommand(load.Cmd)
 	RootCmd.AddCommand(mongoload.Cmd)
-	RootCmd.AddCommand(caliperload.Cmd)
+	RootCmd.AddCommand(jsonload.Cmd)
 	RootCmd.AddCommand(query.Cmd)
 	RootCmd.AddCommand(erclient.Cmd)
 	RootCmd.AddCommand(rdf.Cmd)
